# api/openapi.yaml
openapi: 3.1.0
info:
  title: RatiteRun API
  version: "1.0.0"
  description: |
    REST API приложения RatiteRun — планировщик содержания страусоподобных птиц
    (страус, эму, нанду).

    Ключевые решения:

    * **Стадо — агрегат.** Секции (`housing`, `fencing`, `feed`, …) адресуются
      как самостоятельные подресурсы, потому что каждая соответствует одному
      экрану приложения. Экран редактирует загон — уходит `PUT /housing`,
      а не перезапись всего стада.
    * **Оптимистичные блокировки.** Полный агрегат отдаётся с `ETag` (это
      целочисленная версия). Записи принимают `If-Match`; расхождение — `412`.
    * **Движки оценки на сервере и на клиенте.** Приложение считает готовность
      локально для мгновенного отклика, сервер — для отчётов и веб-доступа.
      Формулы обязаны совпадать: `Engines.php` — порт `Engines.swift`.
    * **Нормативы по видам в базе, а не в бинарнике.** Правка рекомендованной
      площади на птицу не требует релиза в App Store.
    * **Анонимный вход.** Приложение обещает «no account, no sign-up»;
      идентичность привязана к устройству, Apple ID — опциональная надстройка
      для синхронизации между устройствами.

  license:
    name: Proprietary
servers:
  - url: https://api.ratiterun.online/v1
    description: Production
  - url: http://127.0.0.1:8080/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /flocks:
    get:
      tags: [Flocks]
      summary: Список стад (сводки)
      description: |
        Отдаёт сводки, а не полные агрегаты: выбор стада в приложении не должен
        тянуть мегабайты. Готовность уже посчитана и лежит в `readinessPercent`.
      parameters:
        - { name: status,  in: query, schema: { $ref: "#/components/schemas/FlockStatus" } }
        - { name: species, in: query, schema: { $ref: "#/components/schemas/Species" } }
        - { name: limit,   in: query, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { name: cursor,  in: query, schema: { type: string }, description: Из nextCursor предыдущей страницы }
      responses:
        "200":
          description: Страница сводок
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/FlockSummary" }
                  nextCursor:
                    type: [string, "null"]

    post:
      tags: [Flocks]
      summary: Создание стада
      description: |
        `id` можно задать самому: приложение вставляет стадо в список
        оптимистично и не переклеивает идентификатор после ответа.
        Секции заполняются нормативами вида — так же, как это делает
        `AppStore.addEmpty()` на клиенте.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:       { type: string, format: uuid }
                title:    { type: string, maxLength: 160 }
                species:  { $ref: "#/components/schemas/Species" }
                count:    { type: integer, minimum: 1, maximum: 100000 }
                age:      { type: string, maxLength: 80 }
                notes:    { type: string, maxLength: 4000 }
                priority: { $ref: "#/components/schemas/Priority" }
                status:   { $ref: "#/components/schemas/FlockStatus" }
      responses:
        "201":
          description: Создано
          headers:
            ETag:     { $ref: "#/components/headers/ETag" }
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Flock" }
        "409": { description: Стадо с таким id уже существует }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /flocks/{id}/reports:
    parameters:
      - $ref: "#/components/parameters/FlockId"
    get:
      tags: [Reports]
      summary: История отчётов
      responses:
        "200":
          description: Список
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Report" }
    post:
      tags: [Reports]
      summary: Сборка PDF-отчёта
      description: |
        Разделы и формулировки те же, что в `ReportGenerator.swift`.
        `shareable: true` выдаёт публичную ссылку — ветеринар или инспектор
        открывает PDF без аккаунта.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sections:
                  type: array
                  items:
                    type: string
                    enum: [spaceFence, diet, safety, health, breeding]
                notes:     { type: string, maxLength: 4000 }
                currency:  { type: string, maxLength: 8, default: "£" }
                shareable: { type: boolean, default: false }
      responses:
        "201":
          description: Отчёт готов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
  /reports/{reportId}:
    get:
      tags: [Reports]
      summary: Метаданные отчёта
      parameters:
        - { name: reportId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Отчёт
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
...
  /shared/reports/{token}:
    get:
      tags: [Reports]
      summary: Публичный доступ к отчёту
      description: Токен случайный, отчёт протухает по `expiresAt`.
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string, pattern: "^[0-9a-f]{32}$" } }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404": { description: Ссылка истекла или не существует }

  # ----------------------------------------------------------- Reference ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    FlockId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 191 }
      description: |
        Повтор с тем же ключом и тем же телом возвращает сохранённый ответ
        вместо создания дубликата. Тот же ключ с другим телом — `409`.
  headers:
    ETag:
      schema: { type: string, example: '"3"' }
      description: Целочисленная версия ресурса.
  responses:
...
    ValidationError:
      description: Тело не прошло валидацию
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 application/problem+json
      properties:
        type:     { type: string, example: "https://api.ratiterun.online/problems/validation" }
        title:    { type: string, example: Unprocessable Entity }
        status:   { type: integer, example: 422 }
        detail:   { type: string }
        instance: { type: string }
        errors:
          type: object
          description: Пополевые ошибки валидации
          additionalProperties:
            type: array
            items: { type: string }
          example:
            strength: ["Must be between 1 and 5."]
    Species:
      type: string
      enum: [ostrich, emu, rhea]
    Priority:
      type: string
      enum: [low, medium, high]
    FlockStatus:
      type: string
      enum: [setup, active, attention, danger]
...
    ReminderKind:
      type: string
      enum: [feedGrit, water, fenceCheck, legHealth]
    LayoutKind:
      type: string
      enum: [paddock, shelter, feeder, waterer, gate, dustBath]
    TerrainType:
      type: string
      enum: [flat, sloped, sandy, grassland, mixed]
    DietType:
      type: string
      enum: [grazing, formulated, mixed]
...
    Stroke:
      type: array
      description: Один штрих — последовательность точек в нормализованных координатах.
      items:
        type: object
        properties:
          x: { type: number }
          y: { type: number }
    FlockSummary:
      type: object
      properties:
        id:               { type: string, format: uuid }
        title:            { type: string }
        species:          { $ref: "#/components/schemas/Species" }
        count:            { type: integer }
        age:              { type: string }
        priority:         { $ref: "#/components/schemas/Priority" }
        status:           { $ref: "#/components/schemas/FlockStatus" }
        readinessPercent: { type: integer, minimum: 0, maximum: 100 }
        photoUrl:         { type: [string, "null"] }
        version:          { type: integer }
        createdDate:      { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
    Flock:
      allOf:
        - $ref: "#/components/schemas/FlockSummary"
        - type: object
          properties:
            notes:     { type: string }
            housing:   { $ref: "#/components/schemas/Housing" }
            fencing:   { $ref: "#/components/schemas/Fencing" }
            feed:      { $ref: "#/components/schemas/Feed" }
            waterGrit: { $ref: "#/components/schemas/WaterGrit" }
            handling:  { $ref: "#/components/schemas/Handling" }
            breeding:  { $ref: "#/components/schemas/Breeding" }
            rearing:   { $ref: "#/components/schemas/Rearing" }
            health:    { $ref: "#/components/schemas/Health" }
            predator:  { $ref: "#/components/schemas/Predator" }
            terrain:   { $ref: "#/components/schemas/Terrain" }
            kit:       { $ref: "#/components/schemas/MaterialKit" }
            signoff:   { $ref: "#/components/schemas/Signoff" }
            markup:    { $ref: "#/components/schemas/Markup" }
            birds:
              type: array
              items: { $ref: "#/components/schemas/BirdRecord" }
            reminders:
              type: array
              items: { $ref: "#/components/schemas/Reminder" }
            layout:
              type: array
              items: { $ref: "#/components/schemas/LayoutItem" }
            readiness: { $ref: "#/components/schemas/Readiness" }
...
    Housing:
      type: object
      properties:
        spacePerBird: { type: number, description: "м² на птицу" }
        shelterArea:  { type: number, description: "м² укрытия" }
        paddockSize:  { type: number, description: "м² загона" }
        terrain:      { $ref: "#/components/schemas/TerrainType" }
        hasShelter:   { type: boolean }
    Fencing:
      type: object
      properties:
        height:           { type: number, description: "метры" }
        strength:         { type: integer, minimum: 1, maximum: 5 }
        escapeRiskNote:   { type: string }
        perimeterSecured: { type: boolean }
    Feed:
      type: object
      properties:
        dietType:     { $ref: "#/components/schemas/DietType" }
        grazingRatio: { type: number, minimum: 0, maximum: 100 }
        proteinPct:   { type: number, minimum: 0, maximum: 100 }
        scheduleNote: { type: string }
    WaterGrit:
      type: object
      properties:
        waterProvided:    { type: boolean }
        gritProvided:     { type: boolean }
        gritGramsPerBird: { type: number }
        mineralsProvided: { type: boolean }
        notes:            { type: string }
    Handling:
      type: object
      properties:
        neverCorner:         { type: boolean }
        approachFromSide:    { type: boolean }
        useHood:             { type: boolean }
        trainedHandlersOnly: { type: boolean }
        restraintPlan:       { type: string }
    Breeding:
      type: object
      properties:
        pairs:          { type: integer }
        largeEggs:      { type: integer }
        incubationNote: { type: string }
        season:         { type: string }
        startDate:      { type: string, format: date-time }
    Rearing:
      type: object
      properties:
        chicks:           { type: integer }
        brooderReady:     { type: boolean }
        legIssuesFlagged: { type: boolean }
        chickDietNote:    { type: string }
    Health:
      type: object
      properties:
        legJointScore: { type: integer, minimum: 1, maximum: 5 }
        ailmentsNote:  { type: string }
        vetContact:    { type: string }
        lastCheck:     { type: string, format: date-time }
    Predator:
      type: object
      properties:
        fenceIntegrity: { type: integer, minimum: 1, maximum: 5 }
        predatorsSeen:  { type: string }
        securityNote:   { type: string }
        lastChecked:    { type: string, format: date-time }
    Terrain:
      type: object
      properties:
        dustBathing: { type: boolean }
        roomToRun:   { type: boolean }
        terrainNote: { type: string }
    MaterialKit:
      type: object
      description: Считается сервером при каждом изменении площади или поголовья.
      properties:
        fenceMeters:   { type: number }
        feeders:       { type: integer }
        waterers:      { type: integer }
        gritKg:        { type: number }
        shelterM2:     { type: number }
        safetyKitNote: { type: string }
    Signoff:
      type: object
      properties:
        reviewer: { type: string }
        signatureStrokes:
          type: array
          items: { $ref: "#/components/schemas/Stroke" }
        date:     { type: string, format: date-time }
        approved: { type: boolean }
    Markup:
      type: object
      properties:
        strokes:
          type: array
          items: { $ref: "#/components/schemas/Stroke" }
        caption: { type: string }
    BirdRecord:
      type: object
      properties:
        id:       { type: string, format: uuid }
        birdID:   { type: string }
        weightKg: { type: number }
        heightCm: { type: number }
        note:     { type: string }
        date:     { type: string, format: date-time }
    Reminder:
      type: object
      properties:
        id:      { type: string, format: uuid }
        kind:    { $ref: "#/components/schemas/ReminderKind" }
        title:   { type: string }
        hour:    { type: integer }
        minute:  { type: integer }
        enabled: { type: boolean }
    LayoutItem:
      type: object
      properties:
        id:   { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/LayoutKind" }
        x:    { type: number, minimum: 0, maximum: 1 }
        y:    { type: number, minimum: 0, maximum: 1 }
    Readiness:
      type: object
      properties:
        percent: { type: integer, minimum: 0, maximum: 100 }
        band:    { type: string, enum: ["Ready", "Getting there", "Needs work", "Not ready"] }
        parts:
          type: object
          properties:
            space:     { type: number }
            fencing:   { type: number }
            diet:      { type: number }
            gritWater: { type: number }
            safety:    { type: number }
            health:    { type: number }
...
    Report:
      type: object
      properties:
        id:        { type: string, format: uuid }
        flockId:   { type: string, format: uuid }
        status:    { type: string, enum: [pending, ready, failed] }
        sections:
          type: array
          items: { type: string }
        notes:     { type: string }
        currency:  { type: string }
        pdfUrl:    { type: string }
        shareUrl:  { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }